Centralize the blocking/waiting code into the status lib
